isolated the http-client in use by MCRRESTResolver and allow to config-override the used class, consult the added MCRHTTPClient interface to see the simple abstraction
